handle detail page

+++ app/src/main/java/com/gphackathon/data/models/TvSeriesData.kt
@@ -14,7 +14,7 @@ data class TvSeriesData(
     val overview: String,
     val popularity: Double,
     val poster_path: String,
-    val vote_average: Int,
+    val vote_average: Float,
     val vote_count: Int
 ) {
     fun getPosterImageUrl(): String {
